requirements.txt
Django==3.2.7
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
mysqlclient==2.0.3
pandas==1.3.3
dash==2.0.0
dash-bootstrap-components==0.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
django_plotly_dash
dpd-static-support
django-bootstrap4


